Exception handling - delete

+++ src/main/java/com/educandoweb/course/service/UserService.java
@@ -2,8 +2,11 @@ package com.educandoweb.course.service;
 
 import com.educandoweb.course.entities.User;
 import com.educandoweb.course.repository.UserRepository;
+import com.educandoweb.course.service.exceptions.DatabaseException;
 import com.educandoweb.course.service.exceptions.ResourcesNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 
@@ -30,7 +33,16 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)){
+            throw new ResourcesNotFoundException(id);
+        }
+        try {
+            repository.deleteById(id);
+        }catch (EmptyResultDataAccessException e){
+            throw new ResourcesNotFoundException(id);
+        }catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
